type-safer ColumnString::EstimatedValueSize, better validation, more tests
- Added validation for ColumnString::EstimatedValueSize value in ColumnString
- Minor refactoring of block allocation in ColumnString
- Moved test code for ColumnString and ColumnLowCardinality to separate files,
- Added more test cases
