+++ ut/utils.h
@@ -3,6 +3,7 @@
 #include <clickhouse/base/platform.h>
 #include <clickhouse/base/uuid.h>
 
+#include "clickhouse/columns/string.h"
 #include "clickhouse/query.h"
 #include "utils_meta.h"
 #include "utils_comparison.h"
@@ -29,6 +30,8 @@ namespace clickhouse {
     struct Profile;
     struct QuerySettingsField;
     struct Progress;
+
+    using ColumnRef = std::shared_ptr<class Column>;
 }
 
 template <typename ResultType = std::string>
@@ -144,6 +147,10 @@ std::ostream& operator<<(std::ostream & ostr, const Type & type);
 std::ostream & operator<<(std::ostream & ostr, const ServerInfo & server_info);
 std::ostream & operator<<(std::ostream & ostr, const Profile & profile);
 std::ostream & operator<<(std::ostream & ostr, const Progress & progress);
+inline std::ostream & operator<<(std::ostream & ostr, const ColumnString::EstimatedValueSize & estimation) {
+    return ostr << "ColumnString::EstimatedValueSize{ " << static_cast<std::underlying_type_t<ColumnString::EstimatedValueSize>>(estimation) << " }";
+}
+
 }
 
 std::ostream& operator<<(std::ostream & ostr, const PrettyPrintBlock & block);
@@ -207,3 +214,5 @@ inline uint64_t versionNumber(
 uint64_t versionNumber(const clickhouse::ServerInfo & server_info);
 
 std::string ToString(const clickhouse::UUID& v);
+
+std::ostream & dumpMemoryUsage(const char * prefix, const clickhouse::ColumnRef col);
