+++ ut/value_generators.cpp
@@ -29,25 +29,26 @@ std::vector<std::string> MakeFixedStrings(size_t string_size) {
 std::vector<std::string> MakeStrings() {
     return {
         "a", "ab", "abc", "abcd",
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
-        "long string to test how those are handled. Here goes more text. "
+        "A a B b C c D d E e F f G g H h I i J j K k L l M m N n O o P p Q q R r S s T t U u V v W w X x Y y Z z",
+        "line  0: long string to test how those are handled. Here goes more text. \n"
+        "line  1: long string to test how those are handled. Here goes more text. \n"
+        "line  2: long string to test how those are handled. Here goes more text. \n"
+        "line  3: long string to test how those are handled. Here goes more text. \n"
+        "line  4: long string to test how those are handled. Here goes more text. \n"
+        "line  5: long string to test how those are handled. Here goes more text. \n"
+        "line  6: long string to test how those are handled. Here goes more text. \n"
+        "line  7: long string to test how those are handled. Here goes more text. \n"
+        "line  8: long string to test how those are handled. Here goes more text. \n"
+        "line  9: long string to test how those are handled. Here goes more text. \n"
+        "line 10: long string to test how those are handled. Here goes more text. \n"
+        "line 11: long string to test how those are handled. Here goes more text. \n"
+        "line 12: long string to test how those are handled. Here goes more text. \n"
+        "line 13: long string to test how those are handled. Here goes more text. \n"
+        "line 14: long string to test how those are handled. Here goes more text. \n"
+        "line 15: long string to test how those are handled. Here goes more text. \n"
+        "line 16: long string to test how those are handled. Here goes more text. \n"
+        "line 17: long string to test how those are handled. Here goes more text. \n"
+        "line 18: long string to test how those are handled. Here goes more text. "
     };
 }
